refactor(app): split monolithic main.py into dedicated modules
Extract functionality from main.py and utils.py into single-responsibility
modules:
- app.py: MemberApp entry point
- file_io.py: file read/write helpers with error handling
- git_client.py: pygit2 clone/push/branch operations
- github_client.py: PyGitHub API interactions (fork, PR, token auth)
- markdown_builder.py: load/build member markdown files
- strings.py: i18n helper with pybabel
- constants.py: shared platform option tuples and directory constants
- styles.tcss: full Textual CSS stylesheet
- Remove now-empty utils.py and scripts/translate.py
