+++ src/edit_python_pe/styles.tcss
@@ -1,35 +1,149 @@
 /* styles.tcss */
 
-LabeledInput {
+/* Global Colors */
+$gray-0: #ffffff;
+$gray-1: #fafafa;
+$gray-2: #f5f5f5;
+$gray-3: #f0f0f0;
+$gray-4: #d9d9d9;
+$gray-5: #bfbfbf;
+$gray-6: #8c8c8c;
+$gray-7: #595959;
+$gray-8: #434343;
+$gray-9: #262626;
+
+$info: #0958d9;
+$success: #389e0d;
+$warning: #d48806;
+$error: #f5222d;
+
+$bg: $gray-9;
+$fg: $gray-1;
+$bg-light: $gray-1;
+$fg-dark: $gray-9;
+
+/* Textual Core Overrides */
+$primary: $info;
+$background: $bg;
+$foreground: $fg;
+
+Screen {
+    background: $bg;
+    color: $fg;
+}
+
+Button.-primary, .info {
+    background: $info;
+    color: $fg;
+}
+
+Button.-success, .success {
+    background: $success;
+    color: $fg;
+}
+
+Button.-warning, .warning {
+    background: $warning;
+    color: $fg;
+}
+
+Button.-error, .error {
+    background: $error;
+    color: $fg-dark;
+}
+
+.light-bg {
+    background: $bg-light;
+    color: $fg-dark;
+}
+
+FormControl {
     height: auto;
 }
 
-LabeledInput Static {
-    width: 25%;
+FormControl .form-label {
+    width: 100%;
+    text-align: left;
+    color: $gray-4;
+    margin: 1 1;
 }
 
-LabeledInput Input {
+FormControl .form-help-text {
+    width: 100%;
+    text-align: left;
+    color: $gray-5;
+    text-style: italic;
+    padding-left: 1;
+}
+
+FormControl .form-error-text {
+    width: 100%;
+    text-align: left;
+    color: $error;
+    text-style: bold;
+    padding-left: 1;
+    display: none;
+}
+
+FormControl Input, FormControl Select {
     width: 100%;
 }
 
+FormControl TextArea {
+    width: 100%;
+    height: 10;
+}
+
+Input.error, TextArea.error, Select.error {
+    border: tall $error;
+}
+
+FormControl.has-error > Input,
+FormControl.has-error > TextArea,
+FormControl.has-error > Select {
+    border: tall $error;
+}
+
+SocialEntry.has-error .entry-row,
+AliasEntry.has-error .entry-row {
+    border: tall $error;
+}
+
+.entry-row {
+    height: auto;
+    width: 100%;
+}
+
+SocialEntry, AliasEntry {
+    height: auto;
+    margin-bottom: 1;
+}
+
 SocialEntry Select {
-    width: 25%;
+    width: 1fr;
+    margin-left: 1;
+    margin-right: 1;
 }
 
 SocialEntry Input {
-    width: 50%;
+    width: 2fr;
+    margin-right: 1;
 }
 
 SocialEntry Button {
-    width: 25%;
+    width: 1fr;
+    margin-right: 1;
 }
 
 AliasEntry Input {
-    width: 75%;
+    width: 3fr;
+    margin-left: 1;
+    margin-right: 1;
 }
 
 AliasEntry Button {
-    width: 25%;
+    width: 1fr;
+    margin-right: 1;
 }
 
 /* Base styles for screens and structure */
@@ -38,14 +152,35 @@ Screen {
     align: center middle;
 }
 
-#form-container {
+#app-header {
+    dock: top;
+    content-align: center middle;
+    text-style: bold;
+    height: 3;
+    width: 100%;
+}
+
+#app-footer {
+    dock: bottom;
+    content-align: center middle;
+    color: gray;
+    height: 3;
+    width: 100%;
+}
+
+#member-form-scroll-container {
+    height: 1fr;
+    overflow-y: auto;
     padding: 1;
+}
+
+#member-form-social-container, #member-form-alias-container {
+    height: auto;
 }
 
 #auth-container {
     width: 50%;
     height: auto;
-    border: solid green;
     padding: 1;
     align: center middle;
 }
@@ -58,7 +193,6 @@ Screen {
 #lang-container {
     width: 50%;
     height: auto;
-    border: solid blue;
     padding: 1;
     align: center middle;
 }
@@ -68,13 +202,85 @@ Screen {
     margin-bottom: 1;
 }
 
-#lang-actions {
+#member-list-container {
+    width: 50%;
+    height: auto;
+    padding: 1;
+    align: center middle;
+}
+
+#member-list-container Static {
+    text-align: center;
+    margin-bottom: 1;
+}
+
+#lang-actions, #auth-actions, #member-list-actions {
     align: center middle;
     height: auto;
     width: 100%;
 }
 
-#lang-actions Button {
+#member-form-actions {
+    align: center middle;
+    height: auto;
+    width: 100%;
+    padding: 0 1;
+}
+
+#auth-actions-left, #member-list-actions-left, #member-form-actions-left {
+    width: 1fr;
+    height: auto;
+    align: left middle;
+}
+
+#auth-actions-right, #member-list-actions-right, #member-form-actions-right {
+    width: auto;
+    height: auto;
+    align: right middle;
+}
+
+#lang-actions Button, #auth-actions Button, #member-list-actions Button {
+    margin: 1;
+}
+
+#member-form-add-social, #member-form-add-alias {
+    margin: 1 1;
+}
+
+#member-form-actions-left Button {
+    margin: 1 1 1 0;
+}
+
+#member-form-actions-right Button {
+    margin: 1 0 1 1;
+}
+
+DiscardConfirmScreen, QuitConfirmScreen {
+    align: center middle;
+    background: $bg 80%;
+}
+
+#member-form-discard-dialog, #quit-confirm-dialog {
+    width: 50%;
+    height: auto;
+    background: $gray-9;
+    padding: 1;
+    align: center middle;
+    border: solid $primary;
+}
+
+#member-form-discard-dialog-msg, #quit-confirm-dialog-msg {
+    text-align: center;
+    margin-bottom: 1;
+}
+
+#member-form-discard-dialog-actions, #quit-confirm-dialog-actions {
+    align: center middle;
+    width: 100%;
+    height: auto;
+}
+
+#member-form-discard-dialog-actions Button, #quit-confirm-dialog-actions Button {
     margin: 1;
 }
 
@@ -95,11 +301,97 @@ LoadingIndicator {
 }
 
 .subheader {
+    text-align: left;
     text-style: bold;
-    margin-top: 1;
+    margin: 1 1;
+    color: $gray-4;
+}
+
+Input:focus, TextArea:focus {
+    border: tall $info;
+}
+
+Select, Select:focus {
+    border: none;
+    outline: none;
 }
 
 .label {
-    text-align: right;
-    margin-right: 1;
+    text-align: left;
+    width: 100%;
+    margin: 1 1;
+}
+
+/* Loading Screen Styles */
+#loading-container {
+    width: 50%;
+    height: auto;
+    padding: 1;
+    align: center middle;
+}
+
+#loading-container LoadingIndicator {
+    margin-bottom: 1;
+}
+
+#loading-container #loading-msg {
+    text-align: center;
+}
+
+#loading-result-container {
+    align: center middle;
+    height: auto;
+}
+
+#result-icon {
+    text-align: center;
+    margin-bottom: 1;
+}
+
+#result-msg {
+    text-align: center;
+    margin-bottom: 1;
+}
+
+#pr-url {
+    text-align: center;
+    margin-bottom: 1;
+}
+
+.success-icon {
+    color: $success;
+    text-style: bold;
+}
+
+.error-icon {
+    color: $error;
+    text-style: bold;
+}
+
+.error-text {
+    color: $error;
+}
+
+#loading-actions {
+    align: center middle;
+    height: auto;
+    width: 100%;
+    margin-top: 1;
+}
+
+#loading-actions Button {
+    margin: 0 1;
+}
+
+/* Dashboard Screen Styles */
+#dashboard-container {
+    width: 50%;
+    height: auto;
+    padding: 1;
+    align: center middle;
+}
+
+#dashboard-container Button {
+    margin: 1 0;
+    width: 100%;
 }
